Move admin_mode_controller to admin/session_controller
